feat: :sparkles: Challenge 2: Manufacturing Gifts
This commit implements at the manufacturing gifts challenge, which involves  creating a function to determinate which gifts can be crafted with the given materials. the function takes a list of gifts  and string  representing  the available materials  as input,  and returns an array containg the gifts that  can be made  with those materials.

No breaking changes or issues closed by this change.